+++ ex04.cpp
@@ -3,21 +3,21 @@
 using namespace std;
 
 bool tang(string s){
-    for(long long i = 0;i<s.size()-1;i++){
+    for(int i = 0;i<s.size()-1;i++){
         if(s[i]>s[i+1]) return false;
     }
     return true;
 }
 
 bool giam(string s){
-    for(long long i = 0;i<s.size()-1;i++){
+    for(int i = 0;i<s.size()-1;i++){
         if(s[i]<s[i+1]) return false;
     }
     return true;
 }
 
-long long main(){
-    long long t;
+int main(){
+    int t;
     cin>>t;
     while(t--){
         string s;
