+++ ex21.cpp
@@ -18,13 +18,13 @@ string s;
 bool ok = false;
 
 void init(){
-    for(long long i=0;i<s.size();i++){
+    for(int i=0;i<s.size();i++){
         s[i]='0';
     }
 }
 
 void genNext(){
-    long long j = s.size()-1;
+    int j = s.size()-1;
     while(s[j]=='1'){
         s[j]='0';
         j--;
@@ -33,8 +33,8 @@ void genNext(){
     else init();
 }
 
-long long main(){
-    long long t;
+int main(){
+    int t;
     cin>>t;
     while(t--){
         cin>>s;
