+++ ex22.cpp
@@ -13,21 +13,21 @@
 
 using namespace std;
 
-long long a[1005];
-long long n,k;
+int a[1005];
+int n,k;
 
 void init(){
-    for(long long i = 1;i<=k;i++){
+    for(int i = 1;i<=k;i++){
         a[i]=i;
     }
 }
 
 void genNext(){
-    long long i = k;
+    int i = k;
     while(a[i]==n-k+i && i>0) i--;
     if(i>0){
         a[i]++;
-        for(long long j = i+1;j<=k;j++){
+        for(int j = i+1;j<=k;j++){
             a[j]=a[j-1]+1;
         }
     }
@@ -35,7 +35,7 @@ void genNext(){
 }
 
 void display(){
-    for(long long i =1;i<=k;i++){
+    for(int i =1;i<=k;i++){
         cout<<a[i]<<" ";
     }
     cout<<endl;
@@ -43,13 +43,13 @@ void display(){
 
 void input(){
     cin>>n>>k;
-    for(long long i = 1;i<=k;i++){
+    for(int i = 1;i<=k;i++){
         cin>>a[i];
     }
 }
 
-long long main(){
-    long long t;
+int main(){
+    int t;
     cin>>t;
     while(t--){
         memset(a,0,sizeof(a));
