+++ ex24.cpp
@@ -17,13 +17,13 @@ string s;
 bool ok = false;
 
 void init(){
-    for(long long i=0;i<s.size();i++){
+    for(int i=0;i<s.size();i++){
         s[i]='1';
     }
 }
 
 void genNext(){
-    long long j = s.size()-1;
+    int j = s.size()-1;
     while(s[j]=='0'){
         s[j]='1';
         j--;
@@ -34,8 +34,8 @@ void genNext(){
     else init();
 }
 
-long long main(){
-    long long t;
+int main(){
+    int t;
     cin>>t;
     while(t--){
         cin>>s;
