+++ ex26.cpp
@@ -12,35 +12,35 @@
 
 using namespace std;
 
-long long a[1005];
-long long n,k;
+int a[1005];
+int n,k;
 
 void init(){
-    for(long long i = 1;i<=k;i++){
+    for(int i = 1;i<=k;i++){
         a[i]=i;
     }
 }
 
 void genNext(){
-    long long i = k;
+    int i = k;
     while(a[i]==n-k+i && i>0) i--;
     if(i>0){
         a[i]++;
-        for(long long j = i+1;j<=k;j++){
+        for(int j = i+1;j<=k;j++){
             a[j]=a[j-1]+1;
         }
     }
 }
 
 bool isFinal(){
-    for(long long i = 1;i<=k;i++){
+    for(int i = 1;i<=k;i++){
         if(a[i]!=n-k+i) return false;
     }
     return true;
 }
 
 void display(){
-    for(long long i =1;i<=k;i++){
+    for(int i =1;i<=k;i++){
         cout<<a[i];
     }
     cout<<" ";
@@ -50,8 +50,8 @@ void input(){
     cin>>n>>k;
 }
 
-long long main(){
-    long long t;
+int main(){
+    int t;
     cin>>t;
     while(t--){
         memset(a,0,sizeof(a));
