+++ ucln_gcnn.cpp
@@ -14,8 +14,8 @@ long long gcd(long a, long b){
     return gcd(b,a%b);
 }
 
-long long main(){
-    long long t;
+int main(){
+    int t;
     cin>>t;
     while(t--){
         long a,b;
